Address code review
