+++ components/greaselion/browser/DEPS
@@ -4,10 +4,3 @@ include_rules = [
   "+extensions/browser",
   "+extensions/common",
 ]
-
-# Existing exceptions
-specific_include_rules = {
-  "greaselion_service_impl.cc": [
-    "+chrome/browser/extensions/extension_service.h",
-  ],
-}
