+++ components/greaselion/browser/greaselion_service_impl.cc
@@ -221,7 +221,8 @@ GreaselionServiceImpl::GreaselionServiceImpl(
     const base::FilePath& install_directory,
     extensions::ExtensionSystem* extension_system,
     extensions::ExtensionRegistry* extension_registry,
-    scoped_refptr<base::SequencedTaskRunner> task_runner)
+    scoped_refptr<base::SequencedTaskRunner> task_runner,
+    std::unique_ptr<Delegate> delegate)
     : download_service_(download_service),
       install_directory_(install_directory),
       extension_system_(extension_system),
@@ -233,6 +234,7 @@ GreaselionServiceImpl::GreaselionServiceImpl(
       task_runner_(std::move(task_runner)),
       browser_version_(
           version_info::GetBraveVersionWithoutChromiumMajorVersion()),
+      delegate_(std::move(delegate)),
       weak_factory_(this) {
   download_service_->AddObserver(this);
   extension_registry_->AddObserver(this);
@@ -249,12 +251,6 @@ void GreaselionServiceImpl::Shutdown() {
   extension_registry_->RemoveObserver(this);
   task_runner_->PostTask(FROM_HERE,
                          base::BindOnce(&DeleteExtensionDirs, extension_dirs_));
-}
-
-void GreaselionServiceImpl::SetDelegate(std::unique_ptr<Delegate> delegate) {
-  DCHECK(delegate);
-  DCHECK(!delegate_);
-  delegate_ = std::move(delegate);
 }
 
 bool GreaselionServiceImpl::IsGreaselionExtension(const std::string& id) {
@@ -288,7 +284,6 @@ void GreaselionServiceImpl::UpdateInstalledExtensions() {
     // will update the greaselion_extensions_ set. Once it's empty, that
     // callback will call CreateAndInstallExtensions().
     delegate_->UnloadExtension(id);
-    // , extensions::UnloadedExtensionReason::UPDATE
   }
 }
